refactor: split makeOrchestrator from test
